node: factor out kvstore dependency
Hide it behind an interface and inject the dependency via the daemon
initialization.

Signed-off by: Ian Vernon <ian@cilium.io>
