+++ daemon/daemon.go
@@ -72,6 +72,7 @@ import (
 	"github.com/cilium/cilium/pkg/monitor"
 	"github.com/cilium/cilium/pkg/mtu"
 	"github.com/cilium/cilium/pkg/node"
+	nodeStore "github.com/cilium/cilium/pkg/node/store"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy"
 	policyApi "github.com/cilium/cilium/pkg/policy/api"
@@ -1190,8 +1191,9 @@ func NewDaemon() (*Daemon, *endpointRestoreState, error) {
 	// or IPv4 alloc prefix, respectively, retrieved by k8s node annotations.
 	log.Info("Initializing node addressing")
 
-	// Inject BPF dependency into node package.
+	// Inject BPF dependency, kvstore dependency into node package.
 	node.TunnelDatapath = tunnel.TunnelMap
+	node.NodeReg = &nodeStore.NodeRegistrar{}
 
 	if err := node.AutoComplete(); err != nil {
 		log.WithError(err).Fatal("Cannot autocomplete node addresses")
@@ -1325,7 +1327,7 @@ func NewDaemon() (*Daemon, *endpointRestoreState, error) {
 			clustermesh, err := clustermesh.NewClusterMesh(clustermesh.Configuration{
 				Name:            "clustermesh",
 				ConfigDirectory: path,
-				NodeKeyCreator:  node.KeyCreator,
+				NodeKeyCreator:  nodeStore.KeyCreator,
 			})
 			if err != nil {
 				log.WithError(err).Fatal("Unable to initialize ClusterMesh")
